refactor whole package
